2048: Rename to TwentyFourtyEight for GML compatibility
