+++ Userland/Games/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(2048)
 add_subdirectory(BrickGame)
 add_subdirectory(Chess)
 add_subdirectory(ColorLines)
@@ -11,3 +10,4 @@ add_subdirectory(Minesweeper)
 add_subdirectory(Snake)
 add_subdirectory(Solitaire)
 add_subdirectory(Spider)
+add_subdirectory(TwentyFourtyEight)
